+++ src/main/java/com/phantomwing/rusticdelight/food/FoodValues.java
@@ -22,6 +22,8 @@ public class FoodValues {
             .nutrition(1).saturationModifier(0.1F).build();
     public static final FoodProperties ROASTED_COFFEE_BEANS = (new FoodProperties.Builder())
             .nutrition(2).saturationModifier(0.2F).build();
+    public static final FoodProperties GOLDEN_COFFEE_BEANS = (new FoodProperties.Builder())
+            .nutrition(5).saturationModifier(0.8F).build();
     public static final FoodProperties ROASTED_BELL_PEPPER = (new FoodProperties.Builder())
             .nutrition(5).saturationModifier(0.8F).build();
     public static final FoodProperties POTATO_SLICES = (new FoodProperties.Builder())