+++ changelog.md
@@ -1,12 +1,36 @@
 # 1.3.0
 ### Additions
-- Added new Coffee crop
-  - Obtain Coffee Beans by finding Wild Coffee in the jungle or trading with villagers
-- Added drinkable Black Coffee
-  - Drink it to gain a short speed boost
+- Added new Coffee feature. Obtain Coffee Beans by finding Wild Coffee in the jungle or trading with villagers.
+  New items include:
+  - Wild Coffee, a plant found in the jungle that drops Coffee Beans
+  - Coffee Beans, used to grow coffee crops
+  - Roasted Coffee Beans, used to brew coffee
+  - Golden Coffee Beans, used to brew a Potion of Haste
+  - Bag of Coffee Beans
+  - Coffee-Braised Beef
+- Added drinkable Coffee items. Drink them to gain Speed and Haste for a short while. 
+  Some coffee variants have additional effects! New items include:
+  - Coffee
+  - Milk Coffee
+  - Chocolate Coffee
+  - Honey Coffee
+  - Dark Coffee
+- Added Haste potions, including long and strong variants. Brew them with Golden Coffee Beans.
+  - Potion of Haste
+  - Splash Potion of Haste
+  - Lingering Potion of Haste
+  - Arrow of Haste
+- Added new configuration options
+  - Wild Coffee chance, used to configure Wild Coffee spawning. Set to zero to disable the Coffee feature entirely (including trades).
+  - Enable villager trades
+  - Enable wandering trader trades
+  - Enable potions (currently enables brewing a Potion of Haste)
 
 ### Changes
-- Added compatibility with Serene Seasons
+- Improved tooltips to include status effects
+
+### Fixes
+- Fixed compatibility with Serene Seasons
 
 
 # 1.2.1